macOS OpenGL driver doesn't like ints as fragment shader inputs.

+++ src/client/renderer/shaders/sprite_vs.glsl
@@ -33,7 +33,7 @@ out vertex_data {
 	vec3 lightgrid;
 	vec4 color;
 	float lerp;
-	int soft;
+	float soft;
 } vertex;
 
 /**
